+++ ip2geotools_locator/utils.py
@@ -9,3 +9,117 @@ logger = logging.getLogger()
 
 # Location is stored as namedtuple
 Location = namedtuple('Location', 'latitude longitude')
+
+# Default application settings
+DEFAULT_SETTINGS = """
+{
+    "noncommercial": {
+        "ip_city": {
+            "active": true,
+            "generate_marker": true,
+            "api_key": null,
+            "db_file": null,
+            "login": null,
+            "password": null
+        },
+        "host_ip": {
+            "active": true,
+            "generate_marker": true,
+            "api_key": null,
+            "db_file": null,
+            "login": null,
+            "password": null
+        },
+        "ip2location": {
+            "active": false,
+            "generate_marker": false,
+            "api_key": null,
+            "db_file": "",
+            "login": null,
+            "password": null
+        },
+        "ipstack": {
+            "active": false,
+            "generate_marker": false,
+            "api_key": "",
+            "db_file": null,
+            "login": null,
+            "password": null
+        },
+        "max_mind_lite": {
+            "active": false,
+            "generate_marker": false,
+            "api_key": null,
+            "db_file": "",
+            "login": null,
+            "password": null
+        }
+    },
+    "commercial": {
+        "eurek": {
+            "active": false,
+            "generate_marker": false,
+            "api_key": null,
+            "db_file": null,
+            "login": null,
+            "password": null
+        },
+        "geobytes_city": {
+            "active": true,
+            "generate_marker": true,
+            "api_key": null,
+            "db_file": null,
+            "login": null,
+            "password": null
+        },
+        "ip_info": {
+            "active": false,
+            "generate_marker": false,
+            "api_key": "",
+            "db_file": null,
+            "login": null,
+            "password": null
+        },
+        "ip_web": {
+            "active": true,
+            "generate_marker": true,
+            "api_key": null,
+            "db_file": null,
+            "login": null,
+            "password": null
+        },
+        "ip2location_web": {
+            "active": false,
+            "generate_marker": false,
+            "api_key": null,
+            "db_file": null,
+            "login": null,
+            "password": null
+        },
+        "max_mind": {
+            "active": true,
+            "generate_marker": true,
+            "api_key": null,
+            "db_file": null,
+            "login": null,
+            "password": null
+        },
+        "neustar_web": {
+            "active": true,
+            "generate_marker": true,
+            "api_key": null,
+            "db_file": null,
+            "login": null,
+            "password": null
+        },
+        "skyhook": {
+            "active": false,
+            "generate_marker": false,
+            "api_key": null,
+            "db_file": null,
+            "login": null,
+            "password": null
+        }
+    }
+}
+"""